Various improvements
Various improvements, mostly message formatting.

+++ hyperion-buildall.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# Complete SDL-Hercules build using wrljet github mods
+# Complete SDL-Hercules-390 build using wrljet github mods
 # Updated: 25 NOV 2020
 #
 # The most recent version of this script can be ontained with:
@@ -12,54 +12,11 @@
 #
 # Bill Lewis  wrljet@gmail.com
 
-# To run, create a build directory and cd to it, 
-# then run this script.
+#-----------------------------------------------------------------------------
+# To run, create a build directory and cd to it, then run this script.
 #
-#    mkdir herctest && cd herctest
-#    ~/hyperion-buildall.sh 2>&1 | tee ./hyperion-buildall.log
-
-# Show all commands as they are being run
-set -x
-
-# Stop on error
-set -e
-
-TRACE=false
-PROMPTS=false
-
-# Process command line
-POSITIONAL=()
-while [[ $# -gt 0 ]]
-do
-key="$1"
-
-case $key in
-    -t|--trace)
-    TRACE=true
-    shift # past argument
-    ;;
-    -p|--prompts)
-    PROMPTS=true
-    shift # past argument
-    ;;
-    *)    # unknown option
-    POSITIONAL+=("$1") # save it in an array for later
-    shift # past argument
-    ;;
-esac
-done
-set -- "${POSITIONAL[@]}" # restore positional parameters
-
-echo "TRACE       = ${TRACE}"
-echo "PROMPTS     = ${PROMPTS}"
-
-BUILD_DIR=$(pwd)
-
-# Target directory prefix
-TARGET_DIR=$(pwd)/herc4x
-
-echo ${BUILD_DIR}
-echo ${TARGET_DIR}
+#  $ mkdir herctest && cd herctest
+#  $ ~/hercules-helper/hyperion-buildall.sh -v --prompts 2>&1 | tee ./hyperion-buildall.log
 
 #-----------------------------------------------------------------------------
 # git may report:
@@ -83,24 +40,163 @@ echo ${TARGET_DIR}
 git config --global pager.branch false
 
 #-----------------------------------------------------------------------------
+# Stop on error
+set -e
 
-# Print current time and machine cpu type
-date
-uname -m
-gcc --version
-which g++ || true
+# Process command line
+
+if [[ -n $trace ]]  || \
+   [[ -n $TRACE ]]; then
+    set -x # For debugging, show all commands as they are being run
+fi
+
+TRACE=false
+VERBOSE=false
+PROMPTS=false
+
+POSITIONAL=()
+while [[ $# -gt 0 ]]
+do
+key="$1"
+
+case $key in
+    -t|--trace)
+    TRACE=true
+    shift # past argument
+    ;;
+
+    -p|--prompts)
+    PROMPTS=true
+    shift # past argument
+    ;;
+
+    -v|--verbose)
+    VERBOSE=true
+    shift # past argument
+    ;;
+
+    *)    # unknown option
+    POSITIONAL+=("$1") # save it in an array for later
+    shift # past argument
+    ;;
+esac
+done
+set -- "${POSITIONAL[@]}" # restore positional parameters
+
+if [[ "${TRACE}" == true ]]; then
+    # Show all commands as they are being run
+    set -x
+fi
+
+#------------------------------------------------------------------------------
+#                               verbose_msg
+#------------------------------------------------------------------------------
+verbose_msg()
+{
+    if ($VERBOSE); then
+	echo "$1"
+    fi
+}
+
+#------------------------------------------------------------------------------
+#                               detect_system
+#------------------------------------------------------------------------------
+detect_system()
+{
+    echo "System stats:"
+
+    machine=$(uname -m)
+    # echo "Machine is $machine"
+
+    # awk -F= '$1=="ID" { gsub(/"/, "", $2); print $2 ;}' /etc/os-release
+    VERSION_ID=$(awk -F= '$1=="ID" { gsub(/"/, "", $2); print $2 ;}' /etc/os-release)
+    # echo "VERSION_ID is $VERSION_ID"
+
+    VERSION_STR=$(awk -F= '$1=="VERSION_ID" { gsub(/"/, "", $2); print $2 ;}' /etc/os-release)
+    # echo "VERSION_STR is $VERSION_STR"
+
+    verbose_msg "Machine arch     : $machine"
+
+    verbose_msg "Memory Total (MB): $(free -m | awk '/^Mem:/{print $2}')"
+    verbose_msg "Memory Free  (MB): $(free -m | awk '/^Mem:/{print $4}')"
+
+    verbose_msg "VERSION_ID       : $VERSION_ID"
+    verbose_msg "VERSION_STR      : $VERSION_STR"
+
+    # Look for Debian/Ubuntu/Mint
+
+    if [[ $VERSION_ID == debian* || $VERSION_ID == ubuntu* ]]; then
+	# if [[ $(lsb_release -rs) == "18.04" ]]; then
+	VERSION_DISTRO=Debian
+	VERSION_MAJOR=$(echo ${VERSION_STR} | cut -f1 -d.)
+	VERSION_MINOR=$(echo ${VERSION_STR} | cut -f2 -d.)
+
+	echo "OS               : $VERSION_DISTRO variant"
+	echo "OS Version       : $VERSION_MAJOR"
+    fi
+
+    if [[ $VERSION_ID == centos* ]]; then
+	echo "We have a CentOS system"
+
+	# centos-release-7-8.2003.0.el7.centos.x86_64
+	CENTOS_VERS=$(rpm --query centos-release) || true
+	VERSION_MAJOR=$(echo ${CENTOS_VERS#centos-release-} | cut -f1 -d-)
+	VERSION_MINOR=$(echo ${CENTOS_VERS#centos-release-} | cut -f1 -d. | cut -f2 -d-)
+
+	echo "VERSION_MAJOR : $VERSION_MAJOR"
+	if [[ $VERSION_MAJOR -ge 7 ]]; then
+	    echo "CentOS version 7 or later found"
+        fi
+    fi
+}
+
+verbose_msg "Options:"
+verbose_msg "TRACE            : ${TRACE}"
+verbose_msg "VERBOSE          : ${VERBOSE}"
+verbose_msg "PROMPTS          : ${PROMPTS}"
+
+# Detect type of system we're running on and display info
+detect_system
+
+#-----------------------------------------------------------------------------
+# Overall working build diretory is the current director
+BUILD_DIR=$(pwd)
+
+# Prefix (target) directory
+PREFIX_DIR=$(pwd)/herc4x
+
+echo "BUILD_DIR        : ${BUILD_DIR}"
+echo "PREFIX_DIR       : ${PREFIX_DIR}"
+
+#-----------------------------------------------------------------------------
 
 if [ -f /etc/debian_version ]; then
     # Older Debian/Ubuntu/etc.
     OS=Debian
     VER=$(cat /etc/debian_version)
 
-    echo "OS is $OS variant"
-    echo "Version $VER"
+    echo "OS               : $OS variant"
+    echo "OS Version       : $VER"
 fi
 
+echo "Machine arch     : $(uname -m)"
+echo "gcc              : $(gcc --version | head -1)"
+echo "g++ presence     : $(which g++ || true)"
+
+start_seconds="$(TZ=UTC0 printf '%(%s)T\n' '-1')"
+start_time=$(date)
+echo "Processing started: $start_time"
+
+#-----------------------------------------------------------------------------
 # Build Regina Rexx, which we use to run the Hercules tests
-read -p "Hit return to continue (step: build regina rexx)"
+if ($PROMPTS); then
+    read -p "Hit return to continue (Step: Build regina rexx)"
+fi
+
+# Remove any existing Regina, download and untar
+rm -f regina-rexx-3.9.3.tar.gz
+rm -rf regina-rexx-3.9.3/
+
 wget http://www.wrljet.com/ibm360/regina-rexx-3.9.3.tar.gz
 tar xfz regina-rexx-3.9.3.tar.gz 
 cd regina-rexx-3.9.3/
@@ -120,13 +216,17 @@ export CPPFLAGS=-I${BUILD_DIR}/rexx/include
 which rexx
 
 #
-read -p "Hit return to continue (step: hercules clone)"
+if ($PROMPTS); then
+    read -p "Hit return to continue (Step: Hercules clone)"
+fi
+
 cd ${BUILD_DIR}
-mkdir sdl4x
-mkdir ${TARGET_DIR}
+mkdir -p sdl4x
+mkdir -p ${PREFIX_DIR}
 
 # Grab unmodified SDL-Hercules Hyperion repo
 cd sdl4x
+rm -rf hyperion
 git clone https://github.com/SDL-Hercules-390/hyperion.git
 cd hyperion
 git branch -va
@@ -141,10 +241,16 @@ git branch -va
 
 util/bldlvlck 
 
-read -p "Hit return to continue (step: git clone extpkgs)"
+if ($PROMPTS); then
+    read -p "Hit return to continue (Step: git clone extpkgs)"
+fi
+
 cd ${BUILD_DIR}
+rm -rf extpkgs
 mkdir extpkgs
 cd extpkgs/
+
+echo "Cloning extpkgs from https://github.com/wrljet (branch: build-mods-i686)"
 
 git clone -b build-mods-i686 https://github.com/wrljet/gists.git
 cp gists/extpkgs.sh .
@@ -187,7 +293,10 @@ for pgm in "${pgms[@]}"; do
 #   popd > /dev/null;
 done
 
-read -p "Hit return to continue (step: build extpkgs)"
+if ($PROMPTS); then
+    read -p "Hit return to continue (Step: Build extpkgs)"
+fi
+
 cd ${BUILD_DIR}
 cd extpkgs
 
@@ -198,11 +307,14 @@ cd ${BUILD_DIR}/sdl4x/hyperion
 
 # ./autogen.sh
 
-read -p "Hit return to continue (step: configure)"
+if ($PROMPTS); then
+    read -p "Hit return to continue (Step: configure)"
+fi
+
 ./configure \
     --enable-optimization="-O3 -march=native" \
     --enable-extpkgs=${BUILD_DIR}/extpkgs \
-    --prefix=${TARGET_DIR} \
+    --prefix=${PREFIX_DIR} \
     --enable-regina-rexx
 
 # Debian 10 x86_64, gcc 8.3.0
@@ -220,11 +332,17 @@ read -p "Hit return to continue (step: configure)"
 #   --enable-optimization="-O3 -pipe" \
 
 # Compile and link
-read -p "Hit return to continue (step: make)"
+if ($PROMPTS); then
+    read -p "Hit return to continue (Step: make)"
+fi
+
 make clean
 time make -j$(nproc) 2>&1 | tee ${BUILD_DIR}/hyperion-buildall-make.log
 
-read -p "Hit return to continue (step: tests)"
+if ($PROMPTS); then
+    read -p "Hit return to continue (Step: tests)"
+fi 
+
 time make check 2>&1 | tee ${BUILD_DIR}/hyperion-buildall-make-check.log
 # time ./tests/runtest ./tests
 
@@ -243,13 +361,23 @@ time make check 2>&1 | tee ${BUILD_DIR}/hyperion-buildall-make-check.log
 # Quickie test to see if hercules works at all
 # sudo ./hercules
 
-read -p "Hit return to continue (step: install)"
+if ($PROMPTS); then
+    read -p "Hit return to continue (step: install)"
+fi
+
 time make install 2>&1 | tee ${BUILD_DIR}/hyperion-buildall-make-install.log
 
 echo "-----------------------------------------------------------------"
 echo "Done!"
 
+end_time=$(date)
+echo "Processing ended:   $end_time"
 
+elapsed_seconds="$(( $(TZ=UTC0 printf '%(%s)T\n' '-1') - start_seconds ))"
+verbose_msg "total elapsed seconds: $elapsed_seconds"
+echo "Overall elpased time: $( TZ=UTC0 printf '%(%H:%M:%S)T\n' "$elapsed_seconds" )"
+
+#------------------------------------------------------------------------------
 # This last group of helper functions were taken from Fish's extpkgs.sh
 
 #------------------------------------------------------------------------------
